Move vector to separate file, fix filter2
